# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
...
  /players/{playerId}:
    get:
      summary: get all details about a player
      description: useful for player's page
      tags:
        - Players
      parameters:
        - in: path
          name: playerId
          description: player's id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: data about the player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: playerId does not exist
  /players/getPreview/{playerId}:
    get:
      summary: get the preview details of a player
      description: useful for player search
      tags:
        - Players
      parameters:
        - in: path
          name: playerId
          description: player's id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: preview data about the player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPreview"
        "404":
          description: playerId does not exist
  /users/favoritePlayers:
    get:
      tags:
        - Users
        - Players
      summary: Return the favorite players that were saved by user
      responses:
        "200":
          description: players results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPreview"
        "401":
          description: user is not authorized
    post:
      summary: save player in the favorites list of a user
      tags:
        - Users
        - Players
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId:
                  type: integer
                  description: player's id which the user saved
                  example: 85
      responses:
        "201":
          description: The player successfully saved as favorite
        "401":
          description: user is not authorized
  /users/favoriteMatches:
    get:
      tags:
        - Users
        - matches
      summary: Return the favorite matches that were saved by user
      responses:
        "200":
          description: data about the user's favorites matches returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FutureMatch"
        "401":
          description: user is not authorized
    post:
      summary: save match in the favorites list of a user
      tags:
        - Users
        - matches
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                matchId:
                  type: integer
                  description: matches's id which the user saved
                  example: 33
      responses:
        "201":
          description: The match successfully saved as favorite
        "401":
          description: user is not authorized
  /users/UpTo3favoriteMatches:
    get:
      tags:
        - Users
        - matches
      summary: Returns up to 3 favorites matches that were saved by the logged-in user
      parameters:
      responses:
        "200":
          description: data about the user's 3 favorites matches returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PastMatch"
        "401":
          description: user is not authorized
  /users/favoriteTeams:
    get:
      tags:
        - Users
        - Teams
      summary: Return the favorite teams that were saved by user
      responses:
        "200":
          description: data about the user's favorites teams returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: user is not authorized
    post:
      summary: save team in the favorites list of a user
      tags:
        - Users
        - Teams
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId:
                  type: integer
                  description: team's id which the user saved
                  example: 33
      responses:
        "201":
          description: The match successfully saved as favorite
        "401":
          description: user is not authorized
...
  /matches/addMatch:
    post:
      summary: add a match to MATCHES table by body params
      description: adding matches
      tags:
        - matches
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FutureMatch"
      responses:
        "200":
          description: Succesfully added this match
        "404":
          description: Could not add this match due to an error
  /matches/addResult:
    post:
      summary: add a result to a match by body params
      description: adding results to matches
      tags:
        - matches
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                match_id:
                  type: integer
                  description: the match id to add result to
                  example: 23
                match_result:
                  type: string
                  description: the result to add to the match
                  example: 3-2
      responses:
        "200":
          description: Succesfully added result to the match
        "404":
          description: Could not add result to this match due to an error
  /matches/addEventCalendar:
    post:
      summary: add an event calendar to a match by body params
      description: adding event calendars to matches
      tags:
        - matches
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
      responses:
        "200":
          description: Succesfully added event to the match
        "404":
          description: Could not add event to this match due to an error
  /matches/getAllMatches:
    get:
      summary: get all of the matches in the system
      description: returns an array with all the matches in the system
      tags:
        - matches
      responses:
        "200":
          description: Succesfully returned all the matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PastMatch"
        "404":
          description: Could not get matches
  /matches/getAllMatchesSortByDate:
    get:
      summary: get all return all the matches in the system sorted by date
      description: returns an array with all the matches in the system sorted by date
      tags:
        - matches
      responses:
        "200":
          description: Succesfully returned all the matches sorted by date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PastMatch"
        "404":
          description: Could not get matches sorted by date
  /matches/getAllMatchesSortByTeam:
    get:
      summary: get all return all the matches in the system sorted by team
      description: returns an array with all the matches in the system sorted by team
      tags:
        - matches
      responses:
        "200":
          description: Succesfully returned all the matches sorted by team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PastMatch"
        "404":
          description: Could not get matches sorted by team
  /matches/getPastMatches:
    get:
      summary: get all return all the past matches
      description: returns an array with all the past matches in the system
      tags:
        - matches
      responses:
        "200":
          description: Succesfully returned all the past matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PastMatch"
        "404":
          description: Could not get past matches
  /matches/getFutureMatches:
    get:
      summary: get all return all the future matches
      description: returns an array with all the future matches in the system
      tags:
        - matches
      responses:
        "200":
          description: Succesfully returned all the future matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FutureMatch"
        "404":
          description: Could not get future matches
...
  /search/players/{player_query}:
    get:
      summary: search for players on the system
      description: searching players on the system
      tags:
        - search
        - Players
      parameters:
        - in: path
          name: player_query
          description: player's name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: players results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerResult"
        "404":
          description: There seems to be a problem searching players
...
components:
  schemas:
...
    league:
      type: object
      properties:
        league_name:
          type: string
          description: league name
          example: Superliga
        current_season_name:
          type: string
          description: current season name
          example: "2020/2021"
        current_stage_name:
          type: string
          description: current stage name
          example: Championship Round
    Player:
      type: object
      properties:
        player_full_name:
          type: string
          description: player name
          example: James Forrest
        player_team:
          type: string
          description: team name
          example: Celtic
        player_image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/8/172104.png
        player_position:
          type: integer
          description: player's position on the team
        player_common_name:
          type: string
          description: player's common name
          example: J. Forrest
        player_nationality:
          type: string
          description: player's nationality
          example: Scotland
        player_birthdate:
          type: date
          description: player's date of birth
          example: 07/07/1991
        player_height:
          type: string
          description: player's height
          exapmle: 175 cm
        player_weight:
          type: string
          description: player's weight
          example: 71 kg
    PlayerPreview:
      type: object
      properties:
        name:
          type: string
          description: player's name
          example: Sten Michael Grytebust
        image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/11/84587.png
        position:
          type: string
          description: player's position in team
          example: 1
        team_name:
          type: string
          description: player's team name
          example: København
    PlayerResult:
      type: object
      properties:
        player_id:
          type: string
          description: player's id
          example: 5627
        player_full_name:
          type: string
          description: player name
          example: Mikkel Andersen
        player_team_name:
          type: string
          description: team name
          example: Midtjylland
        player_image:
          type: string
          description: player's image
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/27/5627.png
        player_position:
          type: string
          description: player's position on the team
          example: Goalkeeper
...
    Team:
      type: object
      properties:
        name:
          type: string
          description: team name
          example: København
        logo:
          type: string
          description: team logo url
          example: https://cdn.sportmonks.com/images//soccer/teams/21/85.png
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerPreview"
...
    PastMatch:
      type: object
      properties:
        match:
          type: object
          properties:
            match_id:
              type: integer
              description: the match id to add
              example: 23
            match_date:
              type: date
              description: the date of the match
              example: "2021-04-21"
            match_hour:
              type: dateTime
              description: the date and exact hour of the match
              example: "2021-04-21 21:30:00"
            home_team:
              type: string
              description: 3 letters of the home team
              example: FCB
            away_team:
              type: string
              description: 3 letters of the away team
              example: RMA
            referee_id:
              type: integer
              description: id of the match referee
              example: 2
            stadium:
              type: string
              description: the name of the match's stadium
              example: Santiago Bernabeu
            result:
              type: string
              description: the result of the game, left is home team's score and right is away team's score
              example: 3-2
          match_events:
            type: array
            items:
              $ref: "#/components/schemas/league"
    FutureMatch:
      type: object
      properties:
        match_id:
          type: integer
          description: match's id
          example: 123
        match_date:
          type: dateTime
          description: the date of the match
          example: "2021-04-25"
        match_hour:
          type: dateTime
          description: the date and exact hour of the match
          example: "2021-04-25T21:30:00"
        home_team:
          type: string
          description: string of the home team's id
          example: "939"
        away_team:
          type: string
          description: string of the away team's id
          example: "2447"
        referee_id:
          type: integer
          description: referee's id
          example: 1
        stadium:
          type: string
          description: the name of the match's stadium
          example: Santiago Bernabeu
    Event:
      type: object
      properties:
        match_id:
          type: integer
          description: the match id to add result to
          example: 23
        event_id:
          type: integer
          description: the event's id to add to the match
          example: 1
        event_date:
          type: date
          description: the event's date to add to the match
          example: "2021-04-21"
        event_hour:
          type: dateTime
          description: the event's date and time to add to the match
          example: "2021-04-21T21:30:00"
        event_minute:
          type: int
          description: the even'ts minute it happened on the match
          example: 67
        event_description:
          type: string
          description: the event's description
          example: goal